refactor(client): extract palette and theme generation in a custom hook
