+++ packages/client/App.tsx
@@ -1,11 +1,11 @@
 import { DripsyProvider } from 'dripsy';
 import { StatusBar } from 'expo-status-bar';
-import React, { useState } from 'react';
+import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import { colorPalette } from './constants/Colors';
 import { MusicPlayerContextProvider } from './contexts/MusicPlayerContext';
 import useCachedResources from './hooks/useCachedResources';
 import { useSocket } from './hooks/useSocket';
+import { useTheme } from './hooks/useTheme';
 import Navigation from './navigation';
 
 export type SizeTerms = 'xs' | 's' | 'm' | 'l' | 'xl';
@@ -13,44 +13,8 @@ export type BackgroundTerms = 'primary' | 'seconday' | 'white' | 'text';
 
 const App: React.FC = () => {
     const socket = useSocket();
-    const [colorScheme, setColorScheme] = useState<'dark' | 'light'>('dark');
     const isLoadingComplete = useCachedResources();
-    const palette = colorPalette(colorScheme);
-    const theme = {
-        colors: {
-            ...palette,
-        },
-        space: {
-            none: 0,
-            xs: 2,
-            s: 4,
-            m: 8,
-            l: 16,
-            xl: 24,
-        },
-        borderWidths: {
-            s: 1,
-            m: 2,
-            l: 3,
-        },
-        fontSizes: {
-            xs: 14,
-            s: 16,
-            m: 20,
-            l: 24,
-            xl: 32,
-        },
-        radii: {
-            s: 5,
-            m: 10,
-            l: 15,
-            full: 9999,
-        },
-    };
-
-    const toggleColorScheme = () => {
-        setColorScheme(colorScheme === 'dark' ? 'light' : 'dark');
-    };
+    const { colorScheme, theme, toggleColorScheme } = useTheme();
 
     if (!isLoadingComplete) {
         return null;
